+++ blocks/detail-viewer/src/components/StringRenderer/index.tsx
@@ -1,5 +1,5 @@
 import { useBlock } from '@appsemble/preact';
-import { Fragment, h, VNode } from 'preact';
+import { h, VNode } from 'preact';
 
 import type { RendererProps, StringField } from '../../../block';
 
@@ -13,11 +13,11 @@ export default function StringRenderer({ data, field }: RendererProps<StringFiel
   const value = utils.remap(field.name, data);
 
   return (
-    <Fragment>
+    <div className="appsemble-string">
       {label ? <h6 className="title is-6">{label}</h6> : null}
       {value ? (
         <div className="content">{typeof value === 'string' ? value : JSON.stringify(value)}</div>
       ) : null}
-    </Fragment>
+    </div>
   );
 }