+++ packages/app/src/components/FlowPage/index.tsx
@@ -132,7 +132,7 @@ export default function FlowPage({
         data={data}
         ee={ee}
         flowActions={flowActions}
-        prefix={`${prefix}.subPages`}
+        prefix={`${prefix}.subPages.${currentPage}.blocks`}
         showDialog={showDialog}
         transitions
       />