+++ packages/app/src/components/TabsPage/index.tsx
@@ -33,9 +33,13 @@ export default function TabsPage({
         </ul>
       </div>
       <Switch>
-        {subPages.map(({ blocks, name }) => (
+        {subPages.map(({ blocks, name }, index) => (
           <Route key={name} exact path={`${match.path}/${normalize(name)}`}>
-            <BlockList {...blockListProps} blocks={blocks} prefix={`${prefix}.subPages`} />
+            <BlockList
+              {...blockListProps}
+              blocks={blocks}
+              prefix={`${prefix}.subPages.${index}.blocks`}
+            />
           </Route>
         ))}